creds testing/requirements.txt
Flask==2.2.2
azure-identity==1.14.0
azure-keyvault-secrets==4.8.0
cachetools==5.3.0
Werkzeug==2.2.3
